refactor: move `token` view function to enrollment_littlepay
