+++ benefits/enrollment/urls.py
@@ -6,13 +6,14 @@ from django.urls import path
 
 from benefits.routes import routes
 from . import views
+from benefits.enrollment_littlepay.views import token
 
 
 app_name = "enrollment"
 urlpatterns = [
     # /enrollment
     path("", views.index, name=routes.name(routes.ENROLLMENT_INDEX)),
-    path("token", views.token, name=routes.name(routes.ENROLLMENT_TOKEN)),
+    path("token", token, name=routes.name(routes.ENROLLMENT_TOKEN)),
     path("error/reenrollment", views.reenrollment_error, name=routes.name(routes.ENROLLMENT_REENROLLMENT_ERROR)),
     path("retry", views.retry, name=routes.name(routes.ENROLLMENT_RETRY)),
     path("success", views.success, name=routes.name(routes.ENROLLMENT_SUCCESS)),
