+++ benefits/enrollment/views.py
@@ -5,7 +5,6 @@ The enrollment application: view definitions for the benefits enrollment flow.
 import logging
 
 
-from django.http import JsonResponse
 from django.template.response import TemplateResponse
 from django.urls import reverse
 from django.utils.decorators import decorator_from_middleware
@@ -15,8 +14,6 @@ from benefits.routes import routes
 from benefits.core import models, session
 from benefits.core.middleware import EligibleSessionRequired, FlowSessionRequired, pageview_decorator
 
-from benefits.enrollment_littlepay.enrollment import request_card_tokenization_access
-from benefits.enrollment_littlepay.session import Session as LittlepaySession
 from benefits.enrollment_littlepay.views import IndexView as LittlepayIndexView
 from . import analytics
 from .enrollment import Status
@@ -26,35 +23,6 @@ TEMPLATE_SYSTEM_ERROR = "enrollment/system_error.html"
 
 
 logger = logging.getLogger(__name__)
-
-
-@decorator_from_middleware(EligibleSessionRequired)
-def token(request):
-    """View handler for the enrollment auth token."""
-    session = LittlepaySession(request)
-
-    if not session.access_token_valid():
-        response = request_card_tokenization_access(request)
-
-        if response.status is Status.SUCCESS:
-            session.access_token = response.access_token
-            session.access_token_expiry = response.expires_at
-        elif response.status is Status.SYSTEM_ERROR or response.status is Status.EXCEPTION:
-            logger.debug("Error occurred while requesting access token", exc_info=response.exception)
-            sentry_sdk.capture_exception(response.exception)
-            analytics.failed_access_token_request(request, response.status_code)
-
-            if response.status is Status.SYSTEM_ERROR:
-                redirect = reverse(routes.ENROLLMENT_SYSTEM_ERROR)
-            else:
-                redirect = reverse(routes.SERVER_ERROR)
-
-            data = {"redirect": redirect}
-            return JsonResponse(data)
-
-    data = {"token": session.access_token}
-
-    return JsonResponse(data)
 
 
 @decorator_from_middleware(EligibleSessionRequired)
